Move mem_fn tests to functional module
