+++ libs/functional/tests/unit/mem_fn_void_test.cpp
@@ -1,13 +1,3 @@
-#include <hpx/config.hpp>
-#include <hpx/hpx_init.hpp>
-
-#if defined(HPX_MSVC)
-#pragma warning(disable: 4786)  // identifier truncated in debug info
-#pragma warning(disable: 4710)  // function not inlined
-#pragma warning(disable: 4711)  // function selected for automatic inline expansion
-#pragma warning(disable: 4514)  // unreferenced inline removed
-#endif
-
 // Taken from the Boost.Bind library
 //
 //  mem_fn_void_test.cpp - a test for mem_fn.hpp + void returns
@@ -20,6 +10,15 @@
 // accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 //
+
+#include <hpx/config.hpp>
+
+#if defined(HPX_MSVC)
+#pragma warning(disable: 4786)  // identifier truncated in debug info
+#pragma warning(disable: 4710)  // function not inlined
+#pragma warning(disable: 4711)  // function selected for automatic inline expansion
+#pragma warning(disable: 4514)  // unreferenced inline removed
+#endif
 
 #include <hpx/functional/mem_fn.hpp>
