+++ tests/unit/util/CMakeLists.txt
@@ -10,7 +10,6 @@ set(tests
 
 set(subdirs
     function
-    mem_fn
    )
 
 if(HPX_WITH_DISTRIBUTED_RUNTIME)
